Set up maturin build toolchain for Rust FHIRPath parser
Switch build backend from flit to maturin (PyO3) and scaffold a minimal
Rust extension module (`fhirpathpy._rust`). The Rust `parse` stub raises
NotImplementedError for now; an IMPLEMENTED flag gates activation so the
ANTLR parser remains the active fallback. All 1627 existing tests pass.
