+++ Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "fhirpathpy"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+name = "_rust"
+crate-type = ["cdylib"]
+
+[dependencies]
+pyo3 = { version = "0.25", features = ["extension-module"] }
